attestation: Use chromeos::DBusServiceDaemon to implement the daemon
Switched attestationd to use the base D-Bus service daemon helper
class to simplify the daemon's implementation. The base class also
handles SIGTERM, SIGINT, and SIGHUP signals correctly.

Since attestationd runs in a sandbox, I had to allow it to use
rt_sigprocmask and signalfd4 sys calls to watch for incoming signals.

Finally, fixed an issue introduced with a previous change.
Automatic serialization of D-Bus data doesn't pick up derived types
when serializing data. It is not enough to provide serialization
for, say, google::protobuf::MessageLite and have it work with
attestation::StatsResponse which derives from MessageLite.

It is important to specialize those functions for actual types
since to implicit type conversions are performed (we don't want
to succeed serializing a char (which is not supported by D-Bus)
and have it automatically converted to an int).

BUG=None
TEST=USE=attestation emerge-link platform2
     Also tested on actual devices: link (x86-64), spring (ARM), lumpy (x86)

Change-Id: I9e6e9ccc12f9976ff9748b3d06ad1caeb36264ca
Reviewed-on: https://chromium-review.googlesource.com/217419
Reviewed-by: Nam Nguyen <namnguyen@chromium.org>
Tested-by: Alex Vakulenko <avakulenko@chromium.org>
Commit-Queue: Alex Vakulenko <avakulenko@chromium.org>