trunks: Added delegate arg to Verify, and removed CreateAndLoadRSAKey
CreateAndLoadRSAKey was just a wrapper around CreateRSAKeyPair and
LoadKey. I am removing it as it is unnecessary, and there are no clients
that use this method.
Adding a delegate argument for Verify allows us to use it later for
audit sessions. Moreover future testing code in trunks_client_test will
use this AuthorizationDelegate argument.

BUG=None
TEST=unit and trunks_client --regression_test

Change-Id: Idfbcfc6d9a0b063f850fa9a90936c6e887a3a2cd
Reviewed-on: https://chromium-review.googlesource.com/271701
Tested-by: Utkarsh Sanghi <usanghi@chromium.org>
Reviewed-by: Darren Krahn <dkrahn@chromium.org>
Commit-Queue: Utkarsh Sanghi <usanghi@chromium.org>