trunks: Add an FTDI driver interface
This interface allows to talk to the FTDI devices which in turn allow
to connect to SPI devices.

With this in place the host can communicate with SPI based TPM
devices.

These new C files were copied from
https://github.com/danyocom/libmpsse.git (sha1 ec3c3d2b5ae22), where
they are published under GPL-2, and processed by clang-format using
'--style=Chromium'.

BUG=chromium:498861
TEST=none

Change-Id: I700019c144bb702ae765d716a62e3b2e6722d35b
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/275440
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>