tpm_manager: Implemented server/client logic for a few initial commands.
The first commands supported are GetTpmStatus and TakeOwnership. This CL
implements the relevant server and client boilerplate and logic as well
as adds unit tests for this and current untested code.

This CL only includes mock implementations of the LocalDataStore,
TpmStatus, and TpmInitializer interfaces.

BUG=brillo:1040,brillo:1042
TEST=unit

Change-Id: Ie69d343c21ab8abee84eae5c1f3f08f0c37bbeca
Reviewed-on: https://chromium-review.googlesource.com/288367
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>
Commit-Queue: Darren Krahn <dkrahn@chromium.org>
Tested-by: Darren Krahn <dkrahn@chromium.org>