trunks: ftdi: simplify interface management
The FTDI interface driver does not fail immediately after it discovers
a compatible device, but fails to connect to it. It keeps the device
descriptor around, hoping to find a better device, but returning the
failed device pointer if there is no operational one discovered.

This could be useful in other modes of operation, but makes no sense
in the FTDI to SPI driver scenario. This patch makes the driver
stricter: do not keep the device handle in case the device fails to
open.

Not keeping around partially initialized devices allows to make the
code more straightforward.

BUG=chromium:498861
TEST=with the rest of the patches applies verified that SPI over FTDI
     is operational

Change-Id: I54d6812e0a8d1b161d92b9bfae11ff6ad42fc186
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/277162
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>