trunks: ftdi: add TPM_STS register accessors
TPM_STS_X (where X is the locality) needs to be accessed quite
frequently during TPM operation. This patch adds two accessor
functions to read and write that register.

Some code is also added to read this register at startup to partially
verify its validity (the TPM family field is expected to be set to
binary 01, which indicates that this is a TPM2 device), and save the
max burst field's value for future use.

BUG=chromium:498861
TEST=device initialization still succeeds.

Change-Id: Iff3dad6f434c28f82c584862877ca32fc21dbb80
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/277316
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>