trunks: improve FIFO read processing
When reading the FIFO contents in FTDI mode, the host polls the status
register to find out if there is more data available and keeps reading
one byte of FIFO data at a time. This makes FIFO read transactions
very slow.

The typical TPM response always includes a 6 byte header and then some
more data. The actual size of the response is stored in the network
byte order in the last 4 bytes of the header.

To speed up FIFO reads, we now always start with reading the first six
bytes. Once they are read, the total response size can be determined,
and the rest of the payload less one byte can be read. At this point
the 'data available' bit in the TPM_STS register must still be
asserted. Then we read the last byte from the FIFO and verify that the
'data available' is deasserted after that. This verifies that the FIFO
contents and the TPM__STS registers are in sync.

BUG=chrome-os-partner:43025
TEST=verified that trunksd starts up successfully when initializing
     CR50 over the FTDI interface.

Change-Id: I5d9ccf0df94eb7935667feff9e9a00caf447a5ff
Signed-off-by: Vadim Bendebury <vbendeb@chromium.org>
Reviewed-on: https://chromium-review.googlesource.com/293864
Reviewed-by: Utkarsh Sanghi <usanghi@chromium.org>