trunks: Do not re-initialize DBus bus.
Before this CL, trunks_proxy reinitializes the DBus bus on every new
thread it is called. This can lead to memory leaks, and since trunks
hold tpm communication state, it results in errors when called in
multiple threads. This CL enfores the requriment that trunks_proxy must
be intialized and called on the same thread.

BUG=None
TEST=call trunks calls in two threads and observe errors.

Change-Id: I52b714f5c3563bebf6632718a1ef29ef3f7fc0bc