attestation: Fix permissions.
In order to update the attestation database, the attestation user needs
access to /mnt/stateful_partition/unencrypted/preserve via the
'preserve' group. It also needs access to tcsd via the 'tss' group. This
CL sets ownership and permissions on the preserve directory and also
includes all group permissions of the attestation user when dropping
root.

BUG=chrome-os-partner:46244
TEST=manual

Change-Id: I53aab61bc046173e6494fa6876761750c7808c17